interfaces/types,interfaces/builtin,daemon: rename types to builtin
This branch renames the "interfaces/types" package to
"interfaces/builtin" that sounds a lot better and actually makes sense
since the concept of a skill type was replaced with an interface.

The only call site, daemon, is patched to make everything work the same
as before.
